feat: implemented extension to present KPaywallViewController
