+++ Sources/KovaleePaywall/PaywallViewController.swift
@@ -2,6 +2,22 @@ import KovaleeSDK
 import SuperwallKit
 import UIKit
 
+/// ``KPaywallViewController`` is a UIViewController subclass that manages the presentation of a paywall in a full-screen format.
+///
+/// This controller is responsible for initializing with necessary parameters, presenting, and managing a Superwall `PaywallViewController`. It includes delegate handling for paywall completion events and supports modal presentation.
+///
+/// Use this view controller to present a paywall in response to various triggers, with additional custom parameters and completinion.
+
+/// ```swift
+/// let paywallViewController = KPaywallViewController(
+///		event: "button_click",
+///		source: "onboarding"
+///		params: ["user_id": "12345"],
+///		) { vc in
+///			vc.dismiss(animated: true)
+///		}
+/// present(paywallViewController, animated: true, completion: nil)
+/// ```
 public class KPaywallViewController: UIViewController {
     private var spinner: UIActivityIndicatorView?
 
@@ -15,10 +31,17 @@ public class KPaywallViewController: UIViewController {
         self.onComplete(self)
     }
 
+    /// Creates a `KPaywallViewController` instance.
+    ///
+    /// - Parameters:
+    ///   - event: The event trigger for showing the paywall. It refers to the event_name in Superwall.
+    ///   - source: The source from where the paywall has been triggered (ie.  onboarding, home, user profiel etc...). This is useful for tracking purposes.
+    ///   - params: Optional parameters to send to Superwall for filtering audiences.
+    ///   - onComplete: A closure called upon the completion of the paywall interaction. Returns the current ViewController so it can be dismissed.
     public init(
         event: String,
-        params: [String: Any]?,
         source: String,
+        params: [String: Any]?,
         onComplete: @escaping (UIViewController) -> Void
     ) {
         self.event = event
@@ -88,26 +111,5 @@ public class KPaywallViewController: UIViewController {
 
         // Re-enable user interaction
         view.isUserInteractionEnabled = true
-    }
-}
-
-public extension UIViewController {
-    func presentFullScreenPaywallViewController(
-        trigger: String,
-        source: String,
-        params: [String: Any]? = nil,
-        onComplete: (() -> Void)? = nil
-    ) {
-        let paywallViewController = KPaywallViewController(
-            event: trigger,
-            params: params,
-            source: source
-        ) { vc in
-            vc.dismiss(animated: true)
-            onComplete?()
-        }
-
-        paywallViewController.modalPresentationStyle = .fullScreen
-        present(paywallViewController, animated: true, completion: nil)
     }
 }
